Update the Object Model to match DSF 3.5.1: - Added board.inductive_sensor key - Added accelerometer.orientation key - Added fan.tacho_ppr key - Extended the list of supported axis letters (Added g to z) - Modified move_segmentation.segments_per_sec to return float instead of int - Added plugin_manifest.sbc_config_files key - Added dsf.is64bit key - Added offset_adj and slope_adj keys to analog_sensor Added auto_evaluate_expression optional parameter to InterceptConnection class

+++ src/dsf/object_model/move/move_segmentation.py
@@ -5,17 +5,17 @@ class MoveSegmentation(ModelObject):
     """Move segmentation parameters"""
     def __init__(self):
         super().__init__()
-        self._segments_per_sec = 0
-        self._min_segment_length = 0
+        self._segments_per_sec = 0.0
+        self._min_segment_length = 0.0
         
     @property
-    def segments_per_sec(self) -> int:
+    def segments_per_sec(self) -> float:
         """Number of segments per second"""
         return self._segments_per_sec
     
     @segments_per_sec.setter
     def segments_per_sec(self, value):
-        self._segments_per_sec = int(value)
+        self._segments_per_sec = float(value)
         
     @property
     def min_segment_length(self) -> float:
